# docs/api/chat-api-spec.yaml
openapi: 3.1.0
info:
  title: Zentoria Enhanced Chat API
  version: 1.0.0
  description: |
    Backend API for the Enhanced Chat Interface in Zentoria Personal Edition.

    ## Authentication
    All endpoints require authentication via:
    - API Key header: `X-API-Key: znt_test_sk_...`
    - Bearer token: `Authorization: Bearer <jwt>`

    ## Scopes
    Different endpoints require different scopes:
    - `ai.chat` - Chat operations (sessions, messages)
    - `files.read` - Read file attachments
    - `files.write` - Upload file attachments
    - `settings.read` - Read user settings
    - `settings.write` - Modify user settings
    - `admin` - Full access (grants all scopes)
servers:
  - url: https://ai.zentoria.ai/api/v1
    description: Production
  - url: http://localhost:4000/api/v1
    description: Development
security:
  - ApiKeyAuth: []
  - BearerAuth: []
...
paths:
...
  /chat/folders:
    get:
      tags: [Folders]
      operationId: listFolders
      summary: List project folders
      description: Returns all folders for the user in a tree structure.
      parameters:
        - name: flat
          in: query
          schema:
            type: boolean
            default: false
          description: Return flat list instead of tree
      responses:
        '200':
          description: Folder list/tree
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FolderListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
        - ApiKeyAuth: []
        - BearerAuth: []

    post:
      tags: [Folders]
      operationId: createFolder
      summary: Create a folder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFolderRequest'
      responses:
        '201':
          description: Folder created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Folder'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
  /chat/folders/{folderId}:
    get:
      tags: [Folders]
      operationId: getFolder
      summary: Get folder details
      parameters:
        - $ref: '#/components/parameters/FolderIdParam'
      responses:
        '200':
          description: Folder details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FolderWithSessions'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - ApiKeyAuth: []
        - BearerAuth: []

    patch:
      tags: [Folders]
      operationId: updateFolder
      summary: Update folder
      parameters:
        - $ref: '#/components/parameters/FolderIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFolderRequest'
      responses:
        '200':
          description: Folder updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Folder'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - ApiKeyAuth: []
        - BearerAuth: []

    delete:
      tags: [Folders]
      operationId: deleteFolder
      summary: Delete folder
      description: |
        Deletes a folder. Sessions in the folder are moved to unfiled.
        Use `deleteContents=true` to also delete all sessions.
      parameters:
        - $ref: '#/components/parameters/FolderIdParam'
        - name: deleteContents
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '204':
          description: Folder deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
...
  /chat/settings:
    get:
      tags: [Settings]
      operationId: getSettings
      summary: Get user chat settings
      responses:
        '200':
          description: User settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSettings'
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
        - ApiKeyAuth: []
        - BearerAuth: []

    patch:
      tags: [Settings]
      operationId: updateSettings
      summary: Update user chat settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSettingsRequest'
      responses:
        '200':
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSettings'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
        - ApiKeyAuth: []
        - BearerAuth: []

  # ===========================================================================
  # ATTACHMENTS
  # ===========================================================================
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key starting with `znt_test_sk_` or `znt_live_sk_`
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    FolderIdParam:
      name: folderId
      in: path
      required: true
      schema:
        type: string
        pattern: '^fold_[a-zA-Z0-9]{16}$'
...
  schemas:
...
    Session:
      type: object
      required: [id, userId, title, createdAt, updatedAt]
      properties:
        id:
          type: string
          example: sess_abc123def456ghi7
        userId:
          type: string
        title:
          type: string
        folderId:
          type: string
          nullable: true
        agentId:
          type: string
          nullable: true
          description: Default agent for this session
        model:
          type: string
          description: Override model for this session
        systemPrompt:
          type: string
          nullable: true
        messageCount:
          type: integer
        tokenCount:
          type: integer
          description: Total tokens used in session
        isArchived:
          type: boolean
          default: false
        isPinned:
          type: boolean
          default: false
        metadata:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        lastMessageAt:
          type: string
          format: date-time
          nullable: true
...
    Folder:
      type: object
      required: [id, userId, name, createdAt]
      properties:
        id:
          type: string
          example: fold_abc123def456ghi7
        userId:
          type: string
        name:
          type: string
        color:
          type: string
          description: Hex color code
        icon:
          type: string
          description: Icon name or emoji
        parentId:
          type: string
          nullable: true
        position:
          type: integer
          description: Sort order among siblings
        sessionCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    FolderWithSessions:
      allOf:
        - $ref: '#/components/schemas/Folder'
        - type: object
          properties:
            sessions:
              type: array
              items:
                $ref: '#/components/schemas/Session'
            children:
              type: array
              items:
                $ref: '#/components/schemas/Folder'
    FolderListResponse:
      type: object
      properties:
        folders:
          type: array
          items:
            $ref: '#/components/schemas/Folder'
        tree:
          type: array
          items:
            $ref: '#/components/schemas/FolderWithSessions'
          description: Nested folder structure (only if flat=false)
    CreateFolderRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 100
        color:
          type: string
        icon:
          type: string
        parentId:
          type: string
          nullable: true
        position:
          type: integer
    UpdateFolderRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
        color:
          type: string
        icon:
          type: string
        position:
          type: integer

    # -------------------------------------------------------------------------
    # Agents
    # -------------------------------------------------------------------------
...
    UserSettings:
      type: object
      properties:
        defaultAgentId:
          type: string
          nullable: true
        defaultModel:
          type: string
        temperature:
          type: number
        maxTokens:
          type: integer
        streamResponses:
          type: boolean
          default: true
        showTokenCount:
          type: boolean
          default: false
        autoGenerateTitle:
          type: boolean
          default: true
        codeTheme:
          type: string
          enum: [github-dark, github-light, monokai, dracula, one-dark]
        fontSize:
          type: string
          enum: [small, medium, large]
        compactMode:
          type: boolean
          default: false
        showTimestamps:
          type: boolean
          default: true
        enterToSend:
          type: boolean
          default: true
        customSystemPrompt:
          type: string
          nullable: true
        enabledAgents:
          type: array
          items:
            type: string
          description: List of agent IDs user has enabled
    UpdateSettingsRequest:
      type: object
      properties:
        defaultAgentId:
          type: string
          nullable: true
        defaultModel:
          type: string
        temperature:
          type: number
          minimum: 0
          maximum: 2
        maxTokens:
          type: integer
          minimum: 1
          maximum: 16384
        streamResponses:
          type: boolean
        showTokenCount:
          type: boolean
        autoGenerateTitle:
          type: boolean
        codeTheme:
          type: string
        fontSize:
          type: string
        compactMode:
          type: boolean
        showTimestamps:
          type: boolean
        enterToSend:
          type: boolean
        customSystemPrompt:
          type: string
          nullable: true
        enabledAgents:
          type: array
          items:
            type: string

    # -------------------------------------------------------------------------
    # Attachments
    # -------------------------------------------------------------------------
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
            requestId:
              type: string
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...